runtime_service: move warp-sync-tree tests to child module file
